Feat(offer-media): add seller image editor

+++ libs/offer-assets/src/application/use-cases/delete-offer-media.use-case.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { OfferAssetsRepository } from '../../infrastructure/persistence/offer-assets.repository';
 
 @Injectable()
@@ -19,7 +19,23 @@ export class DeleteOfferMediaUseCase {
       throw new NotFoundException('Offer media not found');
     }
 
+    const mediaItems = await this.offerAssetsRepository.findOfferMedia(input.offerId);
+    if (mediaItems.length <= 1) {
+      throw new BadRequestException('Offer must have at least one image');
+    }
+
     await this.offerAssetsRepository.deleteOfferMedia(input.mediaId);
+
+    if (media.mediaType === 'thumbnail') {
+      const nextPrimary = mediaItems.find((item) => item.id !== input.mediaId);
+      if (nextPrimary) {
+        await this.offerAssetsRepository.setOfferPrimaryMedia(
+          input.offerId,
+          nextPrimary.id,
+        );
+      }
+    }
+
     return { deleted: true, id: input.mediaId };
   }
 }
